Enhance Model Selection and Retain User Choices
- Introduced popup for model selection upon CLASSIFY button activation.
- Ensured results are updated post successful classification (200 status from backend).
- Implemented feature to retain user's model selection for ease of subsequent classifications.
- Updated testing documentation to cover new model selection process and user choice retention.

#96 #93
