+++ src/components/body/model_popup/index.tsx
@@ -1,12 +1,19 @@
 import React from "react";
 import { Overlay, InfoContainer } from "./indexElements";
-import { Box, CardHeader, IconButton } from "@mui/material";
+import { Box, CardHeader, IconButton, Button } from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
 import { colours } from "../../../styles/colours";
+import Radio from "@mui/material/Radio";
+import Typography from "@mui/material/Typography";
+import testData from "../../../static_data/static_model_data.json";
 
 interface params {
   setSwitchModelOpen: React.Dispatch<React.SetStateAction<boolean>>;
   switchModelOpen: boolean;
+  selectedModel: string;
+  setSelectedModel: React.Dispatch<React.SetStateAction<string>>;
+  handleInference: () => void;
+  realData: any[]; // Type should be adjusted to match the actual data structure
 }
 
 const SwitchModel: React.FC<params> = (props): JSX.Element => {
@@ -14,12 +21,26 @@ const SwitchModel: React.FC<params> = (props): JSX.Element => {
     props.setSwitchModelOpen(false);
   };
 
+  const selectModel = (model: string): void => {
+    console.log("Model selected:", model);
+    props.setSelectedModel(model);
+  };
+
+  const handleInferenceAndClose = (): void => {
+    props.handleInference();
+    handleClose(); // Call handleClose to close the popup
+  };
+
+  const dataToDisplay =
+    process.env.REACT_APP_MODE === "test" ? testData : props.realData;
+
   return (
     <Overlay>
       <Box
         sx={{
-          width: "15vw",
-          height: "20vh",
+          position: "relative",
+          width: "50vw",
+          height: "65vh",
           zIndex: 30,
           border: `0.01vh solid LightGrey`,
           borderRadius: 1,
@@ -28,7 +49,7 @@ const SwitchModel: React.FC<params> = (props): JSX.Element => {
         boxShadow={1}
       >
         <CardHeader
-          title="Model Information"
+          title="Classification Model Selection"
           titleTypographyProps={{
             variant: "h6",
             align: "left",
@@ -44,7 +65,112 @@ const SwitchModel: React.FC<params> = (props): JSX.Element => {
           }
           sx={{ padding: "0.8vh 0.8vh 0.8vh 0.8vh" }}
         />
-        <InfoContainer></InfoContainer>
+        <InfoContainer>
+          <Typography
+            variant="subtitle1"
+            sx={{ marginTop: 1, marginBottom: 2 }}
+          >
+            Model Selection:
+          </Typography>
+          <Box
+            sx={{
+              display: "grid",
+              gridTemplateColumns: "repeat(2, 1fr)",
+              gap: 1,
+              maxHeight: "40vh",
+              overflowY: "auto",
+              borderBottom: 2,
+              borderColor: "darkgrey",
+            }}
+          >
+            {dataToDisplay.map((data, index) => (
+              <Box
+                key={index}
+                sx={{
+                  display: "flex",
+                  flexDirection: "column",
+                  border: "1px solid lightgrey",
+                  borderRadius: "4px",
+                  padding: "1vh",
+                  cursor: "pointer",
+                  backgroundColor:
+                    props.selectedModel === data.model_name
+                      ? "#f0f0f0"
+                      : "#fff",
+                  "&:hover": {
+                    backgroundColor: "#e0e0e0",
+                  },
+                  width: "34vh",
+                  height: "24vh",
+                  maxWidth: "350px",
+                  maxHeight: "200px",
+                }}
+                onClick={() => {
+                  selectModel(data.model_name);
+                }}
+              >
+                <Box
+                  sx={{
+                    display: "flex",
+                    justifyContent: "space-between",
+                    alignItems: "flex-start",
+                    width: "34vh",
+                    height: "16vh",
+                    maxWidth: "350px",
+                    maxHeight: "200px",
+                  }}
+                >
+                  <Typography fontSize={20} variant="h6">
+                    {data.model_name}
+                  </Typography>
+                  <Radio
+                    checked={props.selectedModel === data.model_name}
+                    onChange={() => {
+                      selectModel(data.model_name);
+                    }}
+                    value={data.model_name}
+                  />
+                </Box>
+                <Typography
+                  variant="body2"
+                  sx={{ fontWeight: "bold", marginBottom: 1 }}
+                >
+                  {data.description}
+                </Typography>
+                <Typography variant="body2" sx={{ marginBottom: 1 }}>
+                  Date: {data.creation_date}
+                </Typography>
+                <Typography variant="body2" sx={{ marginBottom: 1 }}>
+                  Version: {data.version}
+                </Typography>
+                {/* Add more details as needed */}
+              </Box>
+            ))}
+          </Box>
+        </InfoContainer>
+        <Button
+          color="inherit"
+          variant="outlined"
+          onClick={() => {
+            handleInferenceAndClose();
+          }}
+          style={{
+            marginTop: "30px",
+            display: "block",
+            marginLeft: "auto",
+            marginRight: "auto",
+          }}
+        >
+          <div
+            style={{
+              display: "flex",
+              alignItems: "center",
+              flexWrap: "wrap",
+            }}
+          >
+            <span>CLASSIFY</span>
+          </div>
+        </Button>
       </Box>
     </Overlay>
   );
