+++ src/components/body/results_tuner_popup/index.tsx
@@ -3,19 +3,14 @@
 import React from "react";
 import { Overlay, InfoContainer } from "./indexElements";
 import { Box, CardHeader, IconButton, Slider } from "@mui/material";
-import Radio from "@mui/material/Radio";
 import Typography from "@mui/material/Typography";
 import CloseIcon from "@mui/icons-material/Close";
 import { colours } from "../../../styles/colours";
-import testData from "../../../static_data/static_model_data.json";
 
 interface params {
   setResultsTunerOpen: React.Dispatch<React.SetStateAction<boolean>>;
   setScoreThreshold: React.Dispatch<React.SetStateAction<number>>;
   scoreThreshold: number;
-  selectedModel: string;
-  setSelectedModel: React.Dispatch<React.SetStateAction<string>>;
-  realData: any[]; // Type should be adjusted to match the actual data structure
 }
 
 const ResultsTunerPopup: React.FC<params> = (props): JSX.Element => {
@@ -26,14 +21,6 @@ const ResultsTunerPopup: React.FC<params> = (props): JSX.Element => {
   const handleSliderChange = (event: any, value: any): void => {
     props.setScoreThreshold(value);
   };
-
-  const selectModel = (model: string): void => {
-    console.log("Model selected:", model);
-    props.setSelectedModel(model);
-  };
-
-  const dataToDisplay =
-    process.env.REACT_APP_MODE === "test" ? testData : props.realData;
 
   return (
     <Overlay>
@@ -82,85 +69,6 @@ const ResultsTunerPopup: React.FC<params> = (props): JSX.Element => {
             min={10}
             max={90}
           />
-          <Typography
-            variant="subtitle1"
-            sx={{ marginTop: 1, marginBottom: 2 }}
-          >
-            Model Selection:
-          </Typography>
-          <Box
-            sx={{
-              display: "grid",
-              gridTemplateColumns: "repeat(2, 1fr)",
-              gap: 1,
-              maxHeight: "30vh",
-              overflowY: "auto",
-            }}
-          >
-            {dataToDisplay.map((data, index) => (
-              <Box
-                key={index}
-                sx={{
-                  display: "flex",
-                  flexDirection: "column",
-                  border: "1px solid lightgrey",
-                  borderRadius: "4px",
-                  padding: "1vh",
-                  cursor: "pointer",
-                  backgroundColor:
-                    props.selectedModel === data.model_name
-                      ? "#f0f0f0"
-                      : "#fff",
-                  "&:hover": {
-                    backgroundColor: "#e0e0e0",
-                  },
-                  width: "34vh",
-                  height: "16vh",
-                  maxWidth: "350px",
-                  maxHeight: "200px",
-                }}
-                onClick={() => {
-                  selectModel(data.model_name);
-                }}
-              >
-                <Box
-                  sx={{
-                    display: "flex",
-                    justifyContent: "space-between",
-                    alignItems: "flex-start",
-                    width: "34vh",
-                    height: "16vh",
-                    maxWidth: "350px",
-                    maxHeight: "200px",
-                  }}
-                >
-                  <Typography fontSize={20} variant="h6">
-                    {data.model_name}
-                  </Typography>
-                  <Radio
-                    checked={props.selectedModel === data.model_name}
-                    onChange={() => {
-                      selectModel(data.model_name);
-                    }}
-                    value={data.model_name}
-                  />
-                </Box>
-                <Typography
-                  variant="body2"
-                  sx={{ fontWeight: "bold", marginTop: 1, marginBottom: 1 }}
-                >
-                  {data.description}
-                </Typography>
-                <Typography variant="body2" sx={{ marginBottom: 1 }}>
-                  Date: {data.creation_date}
-                </Typography>
-                <Typography variant="body2" sx={{ marginBottom: 1 }}>
-                  Version: {data.version}
-                </Typography>
-                {/* Add more details as needed */}
-              </Box>
-            ))}
-          </Box>
         </InfoContainer>
       </Box>
     </Overlay>
